requirements.txt
requests
pillow
pystray
loguru
pywin32
